make createHash private
